+++ lib/truffle/truffle/cext_structs.rb
@@ -29,6 +29,7 @@ module Truffle::CExt
       raise TypeError, "wrong argument type #{object.class} (expected T_DATA)"
     end
 
+    Primitive.cext_mark_object_on_call_exit(object) unless Truffle::Interop.null?(Primitive.data_holder_get_marker(data_holder))
     Primitive.data_holder_get_data(data_holder)
   end
 
@@ -68,6 +69,7 @@ class Truffle::CExt::RData
   def polyglot_read_member(name)
     case name
     when 'data'
+      Primitive.cext_mark_object_on_call_exit(@object) unless Truffle::Interop.null?(Primitive.data_holder_get_marker(@data_holder))
       Primitive.data_holder_get_data(@data_holder)
     when 'type'
       type
@@ -294,6 +296,7 @@ class Truffle::CExt::RArrayPtr
   end
 
   def polyglot_as_pointer
+    Primitive.cext_mark_object_on_call_exit(@array)
     Primitive.array_store_address(@array)
   end