+++ src/main/java/org/truffleruby/cext/CExtNodes.java
@@ -30,11 +30,11 @@ import org.truffleruby.builtins.CoreModule;
 import org.truffleruby.builtins.Primitive;
 import org.truffleruby.builtins.PrimitiveArrayArgumentsNode;
 import org.truffleruby.builtins.YieldingCoreMethodNode;
-import org.truffleruby.cext.CExtNodesFactory.CallWithCExtLockNodeFactory;
 import org.truffleruby.cext.CExtNodesFactory.StringToNativeNodeGen;
 import org.truffleruby.cext.UnwrapNode.UnwrapCArrayNode;
 import org.truffleruby.core.MarkingService.ExtensionCallStack;
 import org.truffleruby.core.MarkingServiceNodes;
+import org.truffleruby.core.MarkingServiceNodes.RunMarkOnExitNode;;
 import org.truffleruby.core.array.ArrayToObjectArrayNode;
 import org.truffleruby.core.array.ArrayUtils;
 import org.truffleruby.core.array.RubyArray;
@@ -76,6 +76,7 @@ import org.truffleruby.interop.ToJavaStringNode;
 import org.truffleruby.interop.TranslateInteropExceptionNode;
 import org.truffleruby.core.string.ImmutableRubyString;
 import org.truffleruby.language.LexicalScope;
+import org.truffleruby.language.Nil;
 import org.truffleruby.language.RubyBaseNode;
 import org.truffleruby.language.RubyDynamicObject;
 import org.truffleruby.language.RubyGuards;
@@ -108,7 +109,6 @@ import org.truffleruby.language.objects.AllocationTracing;
 import org.truffleruby.language.objects.MetaClassNode;
 import org.truffleruby.language.objects.WriteObjectFieldNode;
 import org.truffleruby.language.supercall.CallSuperMethodNode;
-import org.truffleruby.language.yield.CallBlockNode;
 import org.truffleruby.parser.IdentifierType;
 
 import com.oracle.truffle.api.CompilerDirectives;
@@ -166,19 +166,52 @@ public class CExtNodes {
     @Primitive(name = "call_with_c_mutex_and_frame")
     public abstract static class CallWithCExtLockAndFrameNode extends PrimitiveArrayArgumentsNode {
 
-        @Child protected CallWithCExtLockNode callCextNode = CallWithCExtLockNodeFactory.create(RubyNode.EMPTY_ARRAY);
-
-        @Specialization
+        @Specialization(limit = "getCacheLimit()")
         protected Object callWithCExtLockAndFrame(
-                VirtualFrame frame, Object receiver, RubyArray argsArray, Object specialVariables, Object block) {
-            final ExtensionCallStack extensionStack = getLanguage().getCurrentFiber().extensionCallStack;
+                VirtualFrame frame, Object receiver, RubyArray argsArray, Object specialVariables, Object block,
+                @CachedLibrary("receiver") InteropLibrary receivers,
+                @Cached ArrayToObjectArrayNode arrayToObjectArrayNode,
+                @Cached TranslateInteropExceptionNode translateInteropExceptionNode,
+                @Cached ConditionProfile ownedProfile,
+                @Cached RunMarkOnExitNode runMarksNode) {
+            final ExtensionCallStack extensionStack = getLanguage()
+                    .getCurrentThread()
+                    .getCurrentFiber().extensionCallStack;
             final boolean keywordsGiven = RubyArguments.getDescriptor(frame) instanceof KeywordArgumentsDescriptor;
             extensionStack.push(keywordsGiven, specialVariables, block);
             try {
-                return callCextNode.execute(receiver, argsArray);
+                final Object[] args = arrayToObjectArrayNode.executeToObjectArray(argsArray);
+
+                if (getContext().getOptions().CEXT_LOCK) {
+                    final ReentrantLock lock = getContext().getCExtensionsLock();
+                    boolean owned = ownedProfile.profile(lock.isHeldByCurrentThread());
+
+                    if (!owned) {
+                        MutexOperations.lockInternal(getContext(), lock, this);
+                    }
+                    try {
+                        return InteropNodes.execute(receiver, args, receivers, translateInteropExceptionNode);
+                    } finally {
+                        runMarksNode.execute(extensionStack);
+                        if (!owned) {
+                            MutexOperations.unlockInternal(lock);
+                        }
+                    }
+                } else {
+                    try {
+                        return InteropNodes.execute(receiver, args, receivers, translateInteropExceptionNode);
+                    } finally {
+                        runMarksNode.execute(extensionStack);
+                    }
+                }
+
             } finally {
                 extensionStack.pop();
             }
+        }
+
+        protected int getCacheLimit() {
+            return getLanguage().options.DISPATCH_CACHE;
         }
     }
 
@@ -192,6 +225,7 @@ public class CExtNodes {
                 @Cached ArrayToObjectArrayNode arrayToObjectArrayNode,
                 @Cached TranslateInteropExceptionNode translateInteropExceptionNode,
                 @Cached ConditionProfile ownedProfile,
+                @Cached RunMarkOnExitNode runMarksNode,
                 @Cached UnwrapNode unwrapNode) {
             final ExtensionCallStack extensionStack = getLanguage().getCurrentFiber().extensionCallStack;
             final boolean keywordsGiven = RubyArguments.getDescriptor(frame) instanceof KeywordArgumentsDescriptor;
@@ -210,13 +244,18 @@ public class CExtNodes {
                         return unwrapNode.execute(
                                 InteropNodes.execute(receiver, args, receivers, translateInteropExceptionNode));
                     } finally {
+                        runMarksNode.execute(extensionStack);
                         if (!owned) {
                             MutexOperations.unlockInternal(lock);
                         }
                     }
                 } else {
-                    return unwrapNode
-                            .execute(InteropNodes.execute(receiver, args, receivers, translateInteropExceptionNode));
+                    try {
+                        return unwrapNode.execute(
+                                InteropNodes.execute(receiver, args, receivers, translateInteropExceptionNode));
+                    } finally {
+                        runMarksNode.execute(extensionStack);
+                    }
                 }
 
             } finally {
@@ -358,6 +397,18 @@ public class CExtNodes {
             final Object[] args = unwrapCArrayNode.execute(argv);
             return sendWithoutCExtLock(frame, receiver, method, block, EmptyArgumentsDescriptor.INSTANCE, args,
                     dispatchNode, ownedProfile);
+        }
+    }
+
+    @Primitive(name = "cext_mark_object_on_call_exit")
+    public abstract static class MarkObjectOnCallExit extends PrimitiveArrayArgumentsNode {
+
+        @Specialization
+        protected Object markOnCallExit(Object object,
+                @Cached WrapNode wrapNode,
+                @Cached MarkingServiceNodes.QueueForMarkOnExitNode markOnExitNode) {
+            markOnExitNode.execute(wrapNode.execute(object));
+            return nil;
         }
     }
 
@@ -1677,6 +1728,7 @@ public class CExtNodes {
         protected Object createNewMarkList(RubyDynamicObject object,
                 @CachedLibrary("object") DynamicObjectLibrary objectLibrary) {
             getContext().getMarkingService().startMarking(
+                    getLanguage().getCurrentThread().getCurrentFiber().extensionCallStack,
                     (Object[]) objectLibrary.getOrDefault(object, Layouts.MARKED_OBJECTS_IDENTIFIER, null));
             return nil;
         }
@@ -1692,7 +1744,8 @@ public class CExtNodes {
             ValueWrapper wrappedValue = toWrapperNode.execute(markedObject);
             if (wrappedValue != null) {
                 noExceptionProfile.enter();
-                getContext().getMarkingService().addMark(wrappedValue);
+                getContext().getMarkingService()
+                        .addMark(getLanguage().getCurrentThread().getCurrentFiber().extensionCallStack, wrappedValue);
             }
             // We do nothing here if the handle cannot be resolved. If we are marking an object
             // which is only reachable via weak refs then the handles of objects it is itself
@@ -1729,27 +1782,10 @@ public class CExtNodes {
             writeMarkedNode.execute(
                     structOwner,
                     Layouts.MARKED_OBJECTS_IDENTIFIER,
-                    getContext().getMarkingService().finishMarking());
+                    getContext().getMarkingService()
+                            .finishMarking(getLanguage().getCurrentThread().getCurrentFiber().extensionCallStack));
             return nil;
         }
-    }
-
-    @CoreMethod(names = "define_marker", onSingleton = true, required = 2)
-    public abstract static class CreateMarkerNode extends CoreMethodArrayArgumentsNode {
-
-        @TruffleBoundary
-        @Specialization
-        protected Object createMarker(RubyDynamicObject object, RubyProc marker) {
-            /* The code here has to be a little subtle. The marker must be associated with the object it will act on,
-             * but the lambda must not capture the object (and prevent garbage collection). So the marking function is a
-             * lambda that will take the object as an argument 'o' which will be provided when the marking function is
-             * called by the marking service. */
-            getContext()
-                    .getMarkingService()
-                    .addMarker(object, (o) -> CallBlockNode.getUncached().yield(marker, o));
-            return nil;
-        }
-
     }
 
     @CoreMethod(names = "rb_thread_check_ints", onSingleton = true, required = 0)
@@ -1990,7 +2026,7 @@ public class CExtNodes {
 
         @Specialization
         protected DataHolder create(Object address) {
-            return new DataHolder(address);
+            return new DataHolder(address, Pointer.NULL);
         }
     }
 
@@ -2009,6 +2045,30 @@ public class CExtNodes {
         @Specialization
         protected Object setData(DataHolder data, Object address) {
             data.setPointer(address);
+            return nil;
+        }
+    }
+
+    @Primitive(name = "data_holder_get_marker")
+    public abstract static class DataHolderGetMarker extends PrimitiveArrayArgumentsNode {
+
+        @Specialization
+        protected Object getMarker(DataHolder data) {
+            return data.getMarker();
+        }
+
+        @Specialization
+        protected Object getMarker(Nil data) {
+            return data;
+        }
+    }
+
+    @Primitive(name = "data_holder_set_marker")
+    public abstract static class DataHolderSetMarker extends PrimitiveArrayArgumentsNode {
+
+        @Specialization
+        protected Object setMarker(DataHolder data, Object marker) {
+            data.setMarker(marker);
             return nil;
         }
     }